requirements.txt
aiohttp
autogen-agentchat
autogen
python-dotenv==1.0.0
openai
azure-identity==1.15.0
sse-starlette==1.6.5 